requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
python-dotenv>=1.0
httpx>=0.25
openai>=1.12
Pillow>=10.0
pydantic>=2.0
aiofiles>=23.0

